ch4&5 exercises

+++ python4everyone_examples/Ch2/ex3.py
@@ -2,5 +2,5 @@
 
 hrs = input('Enter Hours:')
 rt = input('Enter rate:')
-pay = int(hrs)*float(rt)
+pay = float(hrs)*float(rt)
 print('Pay:', pay)
